LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/content-security-policy/reporting-api/report-to-directive-allowed-in-meta.https.sub-expected.txt



PASS Test that image does not load
PASS Event is fired
FAIL Report is observable to ReportingObserver assert_equals: expected 54 but got 0
FAIL Violation report status OK. undefined is not an object (evaluating 'data[0]["body"]')
